Add global visibility to pool
