+++ src/main/java/org/jvnet/hudson/plugins/port_allocator/Pool.java
@@ -9,6 +9,7 @@ public class Pool {
     
     public String name;
     public String ports;
+    public boolean isGlobal;
 
     public int[] getPortsAsInt() {
