+++ src/main/webapp/help-pool-definition-visibility.html
@@ -1,0 +1,5 @@
+<div>
+    <p>
+        The visibility of the pool indicating if the allocated port number should be global to the cluster.
+    </p>
+</div>
